% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customInputs-package.R
\docType{package}
\name{customInputs-package}
\alias{customInputs}
\alias{customInputs-package}
\title{Custom Inputs for Shiny}
\description{
customInputs
}
\details{
Custom Inputs for Shiny
}
\author{
Homer White (\email{homerhanumat@gmail.com})
}
\references{
\url{https://github.com/homerhanumat/customInputs}
}
\keyword{package}

