% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{customSliderInput}
\alias{customSliderInput}
\title{Title}
\usage{
customSliderInput(inputId, label, min, max, value, step = NULL,
  round = FALSE, ticks = TRUE, animate = FALSE, width = NULL,
  sep = ",", pre = NULL, post = NULL, timeFormat = NULL,
  timezone = NULL, dragRange = TRUE)
}
\arguments{
\item{dragRange}{}
}
\description{
Title
}

