% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{customTextInput}
\alias{customTextInput}
\title{Custom Text Input}
\usage{
customTextInput(inputId, label, value = "", width = NULL,
  placeholder = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. '400px', or '100\%'.}

\item{placeholder}{A character string giving the user a hint as to what
can be entered into the control. Internet Explorer 8 and 9 do not support
this option.}
}
\description{
Create a custom input control for entry of unstructured
text values.  The control updates if and only if the user presses Enter
or shifts focus away from the control.
}

