% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyCustom.R
\name{useShinyCustom}
\alias{useShinyCustom}
\title{Use the shinyCustom Package}
\usage{
useShinyCustom(slider_policy = "debounce", slider_delay = "250",
  numeric_policy = "debounce", numeric_delay = "250",
  text_policy = "debounce", text_delay = "250")
}
\arguments{
\item{slider_policy}{policy for custom slider}

\item{slider_delay}{delay for cusotm slider}

\item{text_delay}{}
}
\value{
a tag list
}
\description{
Use the shinyCustom Package
}

