% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{customNumericInput}
\alias{customNumericInput}
\title{Title}
\usage{
customNumericInput(inputId, label, value, min = NA, max = NA, step = NA,
  width = NULL)
}
\arguments{
\item{width}{}
}
\description{
Title
}

