% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyCustom.R
\name{useShinyCustom}
\alias{useShinyCustom}
\title{Use the shinyCustom Package}
\usage{
useShinyCustom(slider_policy = "debounce", slider_delay = "250",
  numeric_policy = "debounce", numeric_delay = "250",
  text_policy = "debounce", text_delay = "250", rmd = FALSE,
  html = FALSE)
}
\arguments{
\item{slider_policy}{Policy for a custom slider.  Values are 'debounce' (the
default) and 'throttle'.}

\item{slider_delay}{Delay for a custom slider.  Values are in milliseconds.  The
default is 250 (same as for a regular slider).}

\item{numeric_policy}{Policy for a custom numeric input.  Values are 'debounce'
(the default) and 'throttle'.}

\item{numeric_delay}{Delay for a custom numeric input.  Values are in milliseconds.
The default is 250 (same as for a regular numeric input).}

\item{text_policy}{Policy for a custom text input.  Values are 'debounce'
(the default) and 'throttle'.}

\item{text_delay}{Delay for a custom text input.  Values are in milliseconds.
The default is 250 (same as for a regular text input).}

\item{rmd}{Whether or not the app is an interactive R Mardown document.  Default
is \code{FALSE}.}

\item{html}{Whether or not this is a Shiny app in which the user interface is
built from a custom HTML file.  Default value is \code{FALSE}.}
}
\value{
a tag list
}
\description{
Injects into the Shiny app the JavaScript required to create
custom objects.
}

